pyproject.toml
[project]
name = "mcp-chat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.1.0",
    "langchain-google-genai>=4.0.0",
    "langchain-groq>=1.1.1",
    "langchain-openai>=1.1.4",
    "mcp-use>=1.2.6",
    "mcp>=1.0.0",
    "python-dotenv>=1.0.0",
]
